cpu: Move `intel` submodule to its own file.
