+++ src/cpu.rs
@@ -336,81 +336,4 @@ pub(crate) mod arm {
     }
 }
 
-#[cfg_attr(
-    not(any(target_arch = "x86", target_arch = "x86_64")),
-    allow(dead_code)
-)]
-pub(crate) mod intel {
-    pub(crate) struct Feature {
-        word: usize,
-        mask: u32,
-    }
-
-    impl Feature {
-        #[allow(clippy::needless_return)]
-        #[inline(always)]
-        pub fn available(&self, _: super::Features) -> bool {
-            #[cfg(any(target_arch = "x86", target_arch = "x86_64"))]
-            {
-                prefixed_extern! {
-                    static mut OPENSSL_ia32cap_P: [u32; 4];
-                }
-                return self.mask == self.mask & unsafe { OPENSSL_ia32cap_P[self.word] };
-            }
-
-            #[cfg(not(any(target_arch = "x86", target_arch = "x86_64")))]
-            {
-                return false;
-            }
-        }
-    }
-
-    pub(crate) const FXSR: Feature = Feature {
-        word: 0,
-        mask: 1 << 24,
-    };
-
-    pub(crate) const PCLMULQDQ: Feature = Feature {
-        word: 1,
-        mask: 1 << 1,
-    };
-
-    pub(crate) const SSSE3: Feature = Feature {
-        word: 1,
-        mask: 1 << 9,
-    };
-
-    #[cfg(target_arch = "x86_64")]
-    pub(crate) const SSE41: Feature = Feature {
-        word: 1,
-        mask: 1 << 19,
-    };
-
-    #[cfg(target_arch = "x86_64")]
-    pub(crate) const MOVBE: Feature = Feature {
-        word: 1,
-        mask: 1 << 22,
-    };
-
-    pub(crate) const AES: Feature = Feature {
-        word: 1,
-        mask: 1 << 25,
-    };
-
-    #[cfg(target_arch = "x86_64")]
-    pub(crate) const AVX: Feature = Feature {
-        word: 1,
-        mask: 1 << 28,
-    };
-
-    #[cfg(all(target_arch = "x86_64", test))]
-    mod x86_64_tests {
-        use super::*;
-
-        #[test]
-        fn test_avx_movbe_mask() {
-            // This is the OpenSSL style of testing these bits.
-            assert_eq!((AVX.mask | MOVBE.mask) >> 22, 0x41);
-        }
-    }
-}
+pub mod intel;
